Create merge_TextGrids_from_MFA_with_Original_TextGrids
Merges TextGrid output from the Montreal Forced Aligner (MFA) -- 1 tier for phone boundaries, 1 tier for word boundaries -- with the original TextGrid files (with as many tiers as you want to retain. Here, it is deleting tiers 2 and 3 from the original TextGrid as those are the ones we want to replace. Just change those values (or comment out those lines) if you want to delete (or maintain) different tiers. Saves resulting TextGrid in a new directory.